# docs/bridge-x-server.openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Bridge-X Server",
    "version": "0.0.0",
    "description": "This service has been prepared for getting assets info. Examples of info are status, ip address, name and error code."
  },
  "servers": [
    {
      "url": "https://bridge-server.glory-cloud.dev/api/v1",
      "description": "development cluster in tokyo region"
    },
    {
      "url": "http://localhost:3000",
      "description": "local machine for debug"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/types/{typeId}/assets/{assetId}": {
      "get": {
        "tags": ["Asset"],
        "summary": "get detail information of an asset",
        "description": "This api gets get all information related to a specific asset",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTypeId"
          },
          {
            "$ref": "#/components/parameters/pathAssetId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Asset"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      },
      "put": {
        "tags": ["Asset"],
        "summary": "update asset note",
        "description": "This api update note data of an asset",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTypeId"
          },
          {
            "$ref": "#/components/parameters/pathAssetId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/putAssets"
        },
        "responses": {
          "204": {
            "description": "success"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestBody"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
    "/types/{typeId}/assets/{assetId}/status": {
      "get": {
        "tags": ["Asset Status"],
        "summary": "get status about an asset",
        "description": "This api gets status about an asset.",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTypeId"
          },
          {
            "$ref": "#/components/parameters/pathAssetId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AssetStatus"
          },
          "404": {
            "description": "return 404 when specified asset in path param doesn't exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
...
    "/types/{typeId}/assets/{assetId}/versions": {
      "get": {
        "tags": ["Asset Versions"],
        "summary": "get software versions of the specified asset",
        "description": "get software versions of the specified asset. The path params request typeId and assetId about target asset.",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTypeId"
          },
          {
            "$ref": "#/components/parameters/pathAssetId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AssetVersions"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestPathParams"
          },
          "404": {
            "description": "return 404 when specified asset in path param doesn't exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
    "/packages": {
      "get": {
        "tags": ["Package"],
        "description": "This api gets package list.",
        "parameters": [
          {
            "$ref": "#/components/parameters/queryLimit"
          },
          {
            "$ref": "#/components/parameters/queryOffset"
          },
          {
            "$ref": "#/components/parameters/queryPackageStatus"
          },
          {
            "$ref": "#/components/parameters/queryPackageText"
          },
          {
            "$ref": "#/components/parameters/queryPackageSortName"
          },
          {
            "$ref": "#/components/parameters/querySort"
          }
        ],
        "responses": {
          "200": {
            "headers": {
              "Access-Control-Expose-Headers": {
                "schema": {
                  "type": "string",
                  "default": "X-Total-Count",
                  "description": "The information related to custom headers"
                }
              },
              "X-Total-Count": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$",
                  "description": "total number of assets"
                }
              }
            },
            "$ref": "#/components/responses/Package"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      },
      "post": {
        "tags": ["Package"],
        "description": "create new information related to a specific package",
        "requestBody": {
          "$ref": "#/components/requestBodies/postPackage"
        },
        "responses": {
          "201": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "required": ["packageId", "uploadUrl"],
                  "type": "object",
                  "properties": {
                    "packageId": {
                      "type": "string",
                      "description": "This ID is unique for each package",
                      "example": "00000000-0000-0000-0000-000000000000"
                    },
                    "uploadUrl": {
                      "type": "string",
                      "description": "This is the URL for uploading the package",
                      "example": "https://127.0.0.1/home/xxx/yyy.zip"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestBody"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorised"
          }
        }
      }
    },
    "/packages/{packageId}/uploadCompletion": {
      "put": {
        "tags": ["Package"],
        "description": "When the upload is complete, this API is executed and the status changes from Uploading to Validating. After validation, the status changes from Validating to Complete or Invalid.",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathPackageId"
          }
        ],
        "responses": {
          "204": {
            "description": "success"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
...
    "/packages/{packageId}": {
      "put": {
        "tags": ["Package"],
        "description": "This api related information about a packages.",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathPackageId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/putPackage"
        },
        "responses": {
          "204": {
            "description": "success"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestBody"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      },
      "delete": {
        "tags": ["Package"],
        "description": "This api delete information about a packages.",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathPackageId"
          }
        ],
        "responses": {
          "204": {
            "description": "success"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "pathTypeId": {
        "name": "typeId",
        "schema": {
          "type": "string"
        },
        "in": "path",
        "description": "asset type id",
        "required": true
      },
      "pathAssetId": {
        "name": "assetId",
        "schema": {
          "type": "string"
        },
        "in": "path",
        "description": "asset id",
        "required": true
      },
      "pathPackageId": {
        "name": "packageId",
        "schema": {
          "type": "string"
        },
        "in": "path",
        "description": "package id",
        "required": true
      },
...
      "queryLimit": {
        "name": "limit",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 20
        },
        "in": "query",
        "description": "The number of items to return"
      },
      "queryOffset": {
        "name": "offset",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        },
        "in": "query",
        "description": "The number of items to skip before starting to collect the result set"
      },
...
      "queryPackageSortName": {
        "name": "sortName",
        "schema": {
          "type": "string",
          "description": "possible sorting items.",
          "default": "date",
          "enum": ["name", "status", "summary", "date"]
        },
        "in": "query",
        "description": "Item to sort."
      },
...
      "querySort": {
        "name": "sort",
        "schema": {
          "type": "string",
          "description": "possible sorting order",
          "pattern": "^asc$|^desc$",
          "default": "desc",
          "enum": ["asc", "desc"]
        },
        "in": "query",
        "description": "The order of items"
      },
...
      "queryPackageText": {
        "name": "text",
        "schema": {
          "type": "string"
        },
        "in": "query",
        "description": "free text to seach fields (name, status, uploadBy, summary, description, model, elementsName, elementsVersion, memo). able to use white space for AND search"
      },
...
      "queryPackageStatus": {
        "name": "status",
        "schema": {
          "$ref": "#/components/schemas/EPackageStatus"
        },
        "in": "query"
      },
...
    },
    "requestBodies": {
...
      "putAssets": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "note": {
                  "type": "string",
                  "example": "Teller No.1"
                }
              }
            }
          }
        }
      },
      "postPackage": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "description": "name of the package",
                  "type": "string",
                  "example": "CI-100_1.22"
                }
              },
              "required": ["name"]
            }
          }
        }
      },
      "putPackage": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "memo": {
                  "type": "string",
                  "example": "Package for Glory Bank's CI-100"
                }
              },
              "required": ["memo"]
            }
          }
        }
      },
...
    },
    "responses": {
      "Asset": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AssetInfo"
            }
          }
        }
      },
      "AssetStatus": {
        "description": "return status of an asset or a composite asset. In case that the specified asset is an asset or composite asset hasn't had sub asset, subAssets is empty array.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AssetStatus"
            }
          }
        }
      },
...
      "Package": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Package"
              }
            }
          }
        }
      },
...
      "AssetVersions": {
        "description": "return version of specified asset. when the asset exists and version info is not in system, versions is empty array.",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AssetVersions"
              }
            }
          }
        }
      },
...
      "InvalidRequestBody": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Invalid Request Body"
                }
              }
            }
          }
        }
      },
      "InvalidRequestPathParams": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Invalid Request Path Params"
                }
              }
            }
          }
        }
      },
      "Unauthorised": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorised"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not Found"
                }
              }
            }
          }
        }
      },
...
      "UnexpectedError": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Unexpected Error"
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "AssetStatus": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "string",
            "example": "CI-10"
          },
          "assetId": {
            "type": "string",
            "example": "000001"
          },
          "status": {
            "$ref": "#/components/schemas/EStatus"
          },
          "errorCode": {
            "type": "string",
            "example": "E0001"
          },
          "errorMessage": {
            "type": "string",
            "example": "Something happened in this asset"
          },
          "subAssets": {
            "type": "array",
            "description": "This property has sub asset infomation. when parent asset doesn't have sub asset, this is empty array",
            "items": {
              "$ref": "#/components/schemas/SubAssetStatus"
            }
          }
        },
        "required": ["typeId", "assetId", "status", "subAssets"]
      },
      "SubAssetStatus": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "string",
            "example": "RBW-100"
          },
          "assetId": {
            "type": "string",
            "example": "000002"
          },
          "status": {
            "$ref": "#/components/schemas/EStatus"
          },
          "errorCode": {
            "type": "string",
            "example": "3B27"
          },
          "errorMessage": {
            "type": "string",
            "example": "Sensor error (UV sensor)"
          }
        },
        "required": ["typeId", "assetId", "status"]
      },
      "AssetInfo": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "string",
            "example": "CI-10"
          },
          "assetId": {
            "type": "string",
            "example": "000001"
          },
          "status": {
            "$ref": "#/components/schemas/EStatus"
          },
          "ipAddress": {
            "type": "string",
            "example": "192.168.0.1"
          },
          "note": {
            "type": "string",
            "example": "something like a note"
          },
          "customerId": {
            "type": "string",
            "example": "Glory Global Solution"
          },
          "locationId": {
            "type": "string",
            "example": "Himeji HQ"
          },
          "regionId": {
            "type": "string",
            "example": "East Asia"
          },
          "description": {
            "type": "string",
            "example": "something like a description"
          },
          "alias": {
            "type": "string",
            "example": "alias of asset"
          },
          "installationDate": {
            "type": "string",
            "description": "The timestamp when the asset was installed. This is expressed by ISO 8601 format (YYYY-MM-DDThh:mm:ss[.SSS]+00:00)",
            "format": "date-time",
            "example": "2018-10-14T05:36:56.000Z"
          }
        },
        "required": ["typeId", "assetId", "status"]
      },
...
      "AssetVersions": {
        "description": "The name and the version are about software information of the specified asset's composite devices.",
        "type": "object",
        "properties": {
          "typeId": {
            "type": "string",
            "description": "asset type id",
            "example": "CI-100"
          },
          "assetId": {
            "type": "string",
            "description": "asset id",
            "example": "123456"
          },
          "versions": {
            "type": "array",
            "description": "This property has a array structure about this AssetVersions",
            "items": {
              "$ref": "#/components/schemas/Firmware"
            }
          }
        },
        "example": {
          "typeId": "CI-100",
          "assetId": "123456",
          "versions": [
            {
              "name": "CV_AP",
              "version": "SC12Z"
            },
            {
              "name": "CV_COUNTRY",
              "version": "02000000-02000000"
            },
            {
              "name": "MAIN_AP",
              "version": "C"
            },
            {
              "name": "SPEC_INFO",
              "version": "00-000000"
            }
          ]
        },
        "required": ["typeId", "assetId"]
      },
...
      "EStatus": {
        "description": "This is a status about how the asset works.Possible values are 'Missing','Good' and 'Error'. 'Missing' represents this asset is not connecting to mqtt broker. 'Good' represents this asset is connecting mqtt broker and processing normally. 'Error'represents this asset is connecting to mqtt broker but not processing normally.",
        "type": "string",
        "enum": ["Missing", "Good", "Error"]
      },
      "EPackageStatus": {
        "description": "This is a status about how the package works.",
        "type": "string",
        "enum": ["Uploading", "Validating", "Complete", "Invalid", "Failure"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}